perception lidar component
