+++ modules/perception/lidar_pointcloudcluster/lidar_pointcloudcluster.dag
@@ -1,0 +1,13 @@
+# Define all coms in DAG streaming.
+    module_config {
+    module_library : "/apollo/bazel-bin/modules/perception/lidar_pointcloudcluster/lidar_pointcloudcluster.so"
+    components {
+        class_name : "lidar_pointcloudcluster"
+        config {
+            name : "lidar_pointcloudcluster"
+            readers {
+                channel: "/diamond/sensor/lidar/front/PointCloud2"
+            }
+        }
+      }
+    }
